Moving redux containers to the right spot
These previously were misplaced in components
