+++ webpack/components/QuestionForm.js
@@ -9,8 +9,8 @@ import { responseSetsProps } from '../prop-types/response_set_props';
 
 import Errors from './Errors';
 import ModalDialog from './ModalDialog';
-import ResponseSetModal from './ResponseSetModal';
-import ResponseSetDragWidget from './ResponseSetDragWidget';
+import ResponseSetModal from '../containers/ResponseSetModal';
+import ResponseSetDragWidget from '../containers/ResponseSetDragWidget';
 import CodedSetTableEditContainer from '../containers/CodedSetTableEditContainer';
