+++ webpack/containers/ResponseSetDragWidget.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { Draggable, Droppable } from './Draggable';
-import SearchResult from './SearchResult';
+import { Draggable, Droppable } from '../components/Draggable';
+import SearchResult from '../components/SearchResult';
 import NestedSearchBar from '../components/NestedSearchBar';
 import currentUserProps from '../prop-types/current_user_props';
 import { fetchSearchResults, fetchMoreSearchResults } from '../actions/search_results_actions';
